add hw9 with edits
