+++ homework_9/src/App/App.css
@@ -6,9 +6,14 @@
   padding-left: 100px;
   padding-right: 100px;
 }
-@media screen and (min-device-width: 320px) and (max-device-width: 992px) {
-  .container {
+
+@media screen and (min-device-width: 320px) and (max-device-width: 992px){
+  .container{
     padding-left: 20px;
     padding-right: 20px;
-}
+  }
+  .main-part{
+    padding-left: 0px;
+    padding-right: 0px;
+  }
 }
