+++ homework_9/src/Pages/Main/main.css
@@ -1,0 +1,102 @@
+.block{
+    display: flex;
+    justify-content: space-between;
+    margin-bottom: 47px;
+    width: 100%;
+}
+.decription{
+    display: flex;
+    flex-direction: column;
+    align-items: flex-start;
+    align-content: flex-start;
+    justify-content: center;
+    font-size: 16px;
+    line-height: 26px;
+    width: 45%;
+}
+.img-block{
+    width: 45%;
+}
+.img-block img{
+    width: 100%;
+}
+.left-list{
+    padding-left: 25px;
+}
+.left-list li{
+    list-style-type:disc ;
+}
+h1{
+    margin-bottom: 30px;
+    font-family: 'Baloo';
+    font-size: 36px;
+    line-height: 57px;
+}
+.left-small-text{
+    margin-bottom: 30px;
+}
+h4{
+    margin-bottom: 30px;
+    font-weight: bold;
+}
+.left-list{
+    margin-bottom: 30px;
+}
+h2{
+    margin-bottom: 18px;
+    font-weight: 600;
+    font-size: 24px;
+}
+p{
+    margin-bottom: 18px;
+    font-size: 16px;
+    line-height: 26px;
+    font-weight: 400;
+}
+.btn-subscribe{
+    width: 170px;
+    height: 45px;
+}
+.btn-apply-second{
+    width: 320px;
+    height: 48px;
+}
+
+@media screen and (min-device-width: 320px) and (max-device-width: 992px){
+    .btn-subscribe{
+        width: 100%;
+        margin-bottom: 25px;
+    }
+    .btn-apply-second{
+        width: 100%;
+        margin-bottom: 24px;
+    }
+    .block{
+        flex-direction: column;
+        margin-bottom: 48px;
+    }
+    .block .decription{
+        width: 100%;  
+    }
+    .img-block{
+        width: 100%;
+    }
+    h1{
+        font-weight: 400;
+        font-size: 28px;
+        line-height: 44px;
+        margin-bottom: 14px;
+    }
+    .top-block .left-small-text{
+        margin-bottom: 14px;
+    }
+    h4{
+        margin-bottom: 14px;
+    }
+    .top-block .left-list{
+        margin-bottom: 14px;
+    }
+    .reverse-block{
+        flex-direction: column-reverse;
+    }
+}
