+++ homework_9/src/components/Button/button.js
@@ -1,0 +1,10 @@
+import React from "react";
+import './button.css'
+
+const Button = ({text, castomClassName}) => {
+    return (
+        <button className={`common-btn ${castomClassName}`}>{text}</button>
+    );
+};
+
+export default Button;
