+++ homework_9/src/components/footer/footer-links/footer-links.js
@@ -11,5 +11,5 @@ const FooterLinks = () => {
             </div>
     );
   };
-  
-  export default FooterLinks;
+
+export default FooterLinks;
