+++ homework_9/src/components/footer/footer-text/footer-text.css
@@ -2,7 +2,6 @@
     display: flex;
     flex-direction: column;
 }
-
 footer .left-el-footer .copy-text{
     margin-bottom: 10px;
 }
