+++ homework_9/src/components/footer/footer-text/footer-text.js
@@ -2,7 +2,6 @@ import React from "react";
 import './footer-text.css';
 
 const FooterText = () => {
-  
     return (
             <div className="left-el-footer">
                 <span className="copy-text">© Simple Art, Inc. 2021</span>
@@ -11,4 +10,4 @@ const FooterText = () => {
     );
   };
   
-  export default FooterText;
+export default FooterText;
