+++ homework_9/src/components/header/header-mob-burger/header-mob-burger.css
@@ -5,7 +5,7 @@
 .menu-icon img{
     width: 100%;
 }
-@media screen and (min-device-width: 320px) and (max-device-width: 992px) {
+@media screen and (min-device-width: 320px) and (max-device-width: 992px){
     .menu-icon{
         display: block;
     }
