+++ homework_9/src/components/header/header-navigation/header-navigation.js
@@ -1,6 +1,7 @@
 import React from "react";
 import './header-navigation.css';
 import { NavLink } from 'react-router-dom'
+import Button from "../../Button/button";
 
 const HeaderNavigation = () => {
     return (
@@ -9,7 +10,10 @@ const HeaderNavigation = () => {
             <li className="nav-item"><NavLink className="nav-item-link" to="/teachers">Teachers</NavLink></li>
             <li className="nav-item"><NavLink className="nav-item-link" to="/teachers">Prices</NavLink></li>
             <li className="nav-item"><NavLink className="nav-item-link" to="/about">About us</NavLink></li>
-            <li className="nav-item"><NavLink className="nav-item-link btn-apply-link" to="/about">Apply now</NavLink></li>
+            <li className="nav-item"><NavLink className="nav-item-link" >
+                    <Button text={'Apply now'} castomClassName='btn-apply-first' />       
+            </NavLink></li>
+            
         </ul>
     );
   };
