--- homework_9/src/components/main/block/btn-apply/btn-apply.css
@@ -1,28 +1,0 @@
-.container .decription .bth-second{
-    padding: 10px 118px;
-}
-.btn-apply .btn-apply-link{
-    text-decoration: none;
-    color: #FFFFFF;
-}
-
-body .container .btn-apply-link{
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    background-color:#59ABB0; 
-    border-radius: 10px;
-    padding: 8px 38px;
-    color: #FFFFFF;
-    font-weight: 600;
-}
-body .container .btn-apply-link:hover{
-    background-color: #60DFE8;
-    font-weight: 600;
-}
-@media screen and (min-device-width: 320px) and (max-device-width: 992px) {
-    .btn-apply{
-        width: 100%;
-        margin-bottom: 24px;
-    }
-}
