--- homework_9/src/components/main/block/btn-apply/btn-apply.js
@@ -1,12 +1,0 @@
-import React from "react";
-import './btn-apply.css'
-
-const BtnApply = () => {
-    return (    
-            <div className="btn-apply">
-                <a href="#" className="btn-apply-link bth-second">Apply now</a>
-            </div>
-    );
-};
-
-export default BtnApply;
