--- homework_9/src/components/main/block/text-list-title/text-list-title.css
@@ -1,9 +1,0 @@
-h4{
-    margin-bottom: 30px;
-    font-weight: bold;
-}
-@media screen and (min-device-width: 320px) and (max-device-width: 992px) {
-    h4{
-        margin-bottom: 14px;
-    }
-}
