--- homework_9/src/components/main/block/text-list-title/text-list-title.js
@@ -1,10 +1,0 @@
-import React from "react";
-import './text-list-title.css';
-
-const TextListTitle = () => {
-    return (
-        <h4>With us you will get:</h4> 
-    );
-};
-
-export default TextListTitle;
