--- homework_9/src/components/main/block/text-list/text-list.css
@@ -1,12 +1,0 @@
-.left-list{
-    padding-left: 25px;
-    margin-bottom: 30px;
-}
-.left-list li{
-    list-style-type: disc ;
-}
-@media screen and (min-device-width: 320px) and (max-device-width: 992px) {
-    .top-block .left-list{
-        margin-bottom: 14px;
-    }
-}
