--- homework_9/src/components/main/block/text-list/text-list.js
@@ -1,15 +1,0 @@
-import React from "react";
-import './text-list.css';
-
-const TextList = () => {
-    return (
-        <ul className="left-list">
-            <li>Lorem ipsum dolor sit amet, consectetur adipiscing elit.</li> 
-            <li>Vestibulum commodo id amet non facilisis in.</li>
-            <li>Est eros, vestibulum cursus luctus luctus aenean molestie.</li> 
-            <li>Augue ultricies non tristique malesuada justo neque mi, pellentesque lacus.</li>
-        </ul>
-    );
-};
-
-export default TextList;
