--- homework_9/src/components/main/block/text-main/text-main.css
@@ -1,8 +1,0 @@
-.left-small-text{
-    margin-bottom: 30px;
-}
-@media screen and (min-device-width: 320px) and (max-device-width: 992px) {
-    .top-block .left-small-text{
-        margin-bottom: 14px;
-    }
-}
