--- homework_9/src/components/main/block/text-main/text-main.js
@@ -1,12 +1,0 @@
-import React from "react";
-import './text-main.css';
-
-const TextMain = () => {
-    return (
-        <span className="left-small-text">Our school has more than 20 years of experience in teaching people who always thought that they 
-        don’t have even a slightest hint of a talent. Some other descriptive stuff that I’m too lazy 
-        to make up. </span>
-    );
-};
-
-export default TextMain;
