--- homework_9/src/components/main/block/text-title/text-title.css
@@ -1,14 +1,0 @@
-h1{
-    margin-bottom: 30px;
-    font-family: 'Baloo';
-    font-size: 36px;
-    line-height: 57px;
-}
-@media screen and (min-device-width: 320px) and (max-device-width: 992px) {
-    h1{
-        font-weight: 400;
-        font-size: 28px;
-        line-height: 44px;
-        margin-bottom: 14px;
-    }
-}
