--- homework_9/src/components/main/block/text-title/text-title.js
@@ -1,10 +1,0 @@
-import React from "react";
-import './text-title.css';
-
-const TextTitle = () => {
-    return (
-        <h1>Desperate to learn drawing? We’ll teach you!</h1>
-    );
-};
-
-export default TextTitle;
