--- homework_9/src/components/main/block/top-photo/top-photo.js
@@ -1,12 +1,0 @@
-import React from "react";
-import paris from './paris.svg';
-
-const TopPhoto = () => {
-    return (
-        <div className="img-block">
-            <img src={paris} alt="Photo of Paris" />
-        </div>
-    );
-};
-
-export default TopPhoto;
