--- homework_9/src/components/main/main.css
@@ -1,40 +1,0 @@
-
-/* bloks */
-.block{
-    display: flex;
-    justify-content: space-between;
-    margin-bottom: 47px;
-    width: 100%;
-}
-.img-block{
-    width: 45%;
-}
-.img-block img{
-    width: 100%;
-}
-
-.decription{
-    display: flex;
-    flex-direction: column;
-    align-items: flex-start;
-    align-content: flex-start;
-    justify-content: center;
-    font-size: 16px;
-    line-height: 26px;
-    width: 45%;
-}
-
-@media screen and (min-device-width: 320px) and (max-device-width: 992px) {
-    .block{
-        flex-direction: column;
-        margin-bottom: 48px;
-    }
-
-    .block .decription{
-        width: 100%;  
-    }
-
-    .img-block{
-        width: 100%;
-    }
-}
