--- homework_9/src/components/main/reverse-block/bottom-photo/bottom-photo.js
@@ -1,12 +1,0 @@
-import React from "react";
-import oldLetter from './old-let.svg';
-
-const BottomPhoto = () => {
-    return (
-        <div className="img-block">
-            <img src={oldLetter} alt="Photo of Letter" />
-        </div>
-    );
-};
-
-export default BottomPhoto;
