--- homework_9/src/components/main/reverse-block/btn-subcribe/btn-subcribe.css
@@ -1,24 +1,0 @@
-.btn-subscribe{
-    width: 170px;
-    height: 45px;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    background-color:#59ABB0; 
-    border-radius: 10px;
-    font-weight: 600;
-}
-.btn-subscribe .btn-subscribe-link{
-    text-decoration: none;
-    color: #FFFFFF;
-}
-.btn-subscribe:hover{
-    background-color: #60DFE8;
-}
-
-@media screen and (min-device-width: 320px) and (max-device-width: 992px) { 
-    .btn-subscribe{
-        width: 100%;
-        margin-bottom: 25px;
-    }
-}
