--- homework_9/src/components/main/reverse-block/btn-subcribe/btn-subcribe.js
@@ -1,11 +1,0 @@
-import React from "react";
-import './btn-subcribe.css';
-const BtnSubcribe = () => {
-    return (
-        <div className="btn-subscribe">
-            <a href="#" className="btn-subscribe-link">Subscribe</a>
-        </div>
-    );
-};
-
-export default BtnSubcribe;
