--- homework_9/src/components/main/reverse-block/reverse-block.css
@@ -1,5 +1,0 @@
-@media screen and (min-device-width: 320px) and (max-device-width: 992px) {
-    .reverse-block{
-        flex-direction: column-reverse;
-    }
-}
