--- homework_9/src/components/main/reverse-block/text-p/text-p.css
@@ -1,6 +1,0 @@
-p{
-    margin-bottom: 18px;
-    font-size: 16px;
-    line-height: 26px;
-    font-weight: 400;
-}
