--- homework_9/src/components/main/reverse-block/text-p/text-p.js
@@ -1,10 +1,0 @@
-import React from "react";
-import './text-p.css';
-
-const TextP = () => {
-    return (
-        <p>Subscribe to our newsletter to stay tuned and get hottest updates and deals</p>
-    );
-};
-
-export default TextP;
