--- homework_9/src/components/main/reverse-block/text-question/text-question.css
@@ -1,5 +1,0 @@
-h2{
-    margin-bottom: 18px;
-    font-weight: 600;
-    font-size: 24px;
-}
