--- homework_9/src/components/main/reverse-block/text-question/text-question.js
@@ -1,10 +1,0 @@
-import React from "react";
-import './text-question.css';
-
-const TextQuestion = () => {
-    return (
-        <h2>Interested but currently have other plans?</h2>
-    );
-};
-
-export default TextQuestion;
