+++ homework_9/src/index.js
@@ -1,12 +1,12 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import './index.css';
-import App from './components/app/App';
 import { BrowserRouter } from "react-router-dom";
+import App from './App/App';
 
 ReactDOM.render(
   <BrowserRouter>
-    <App />
+  <App />
   </BrowserRouter>,
   document.getElementById("root")
 );
